Prefab Placement and Win/Lose Conditions
Fixed a bug in the prefabPlacement tool that was not allowing me to select the LayerMask in the custom inspector. Added Win and Lose conditions for the player, When the player reaches the end point, the player wins and if it touches any of the obstacles in the level it dies.