MainMenu and Levels
Created 2 more levels for the game. Each level just has more number of hazards which basically increases the difficulty for the player. Also added a MainMenu which shows up on start-up. NOTE: For the previous commit - Static interactive prefabs were added - when the user picks it up, the user gets a speed boost